+++ test/integration/src/components/contextual/FibonacciTotal/FibonacciTotal.js
@@ -1,17 +1,17 @@
 import React from 'react';
 
 import FibonacciTotalSegment from '../../../segments/FibonacciTotalSegment.js';
-import DataComponent from 'soya/lib/data/redux/DataComponent.js';
-import { SERVER } from 'soya/lib/data/RenderType.js';
+import convert from 'soya/lib/data/redux/convert';
+import { SERVER } from 'soya/lib/data/RenderType';
 
 import style from './style.css';
 
-export default class FibonacciTotal extends DataComponent {
+class FibonacciTotal {
   static getSegmentDependencies() {
     return [FibonacciTotalSegment];
   }
 
-  subscribeQueries(nextProps) {
+  static subscribeQueries(nextProps, subscribe) {
     var hydrationOption = null;
     if (nextProps.loadAtClient) {
       hydrationOption = {
@@ -20,13 +20,12 @@ export default class FibonacciTotal extends DataComponent {
     }
 
     var query = { number: nextProps.number };
-    this.subscribe(FibonacciTotalSegment.id(), query, 'fib',
-      hydrationOption);
+    subscribe(FibonacciTotalSegment.id(), query, 'fib', hydrationOption);
   }
 
   render() {
     var title = `Fibonacci Total (${this.props.number})`;
-    if (!this.state.fib.loaded) {
+    if (!this.props.result.fib.loaded) {
       return <div className={style.container}>
         <h3>{title}</h3>
         <p>Loading...</p>
@@ -34,7 +33,9 @@ export default class FibonacciTotal extends DataComponent {
     }
     return <div className={style.container}>
       <h3>{title}</h3>
-      <p>{this.state.fib.data}</p>
+      <p>{this.props.result.fib.data}</p>
     </div>
   }
 }
+
+export default convert(FibonacciTotal);